Move default fix to Stack generation
